back-flask/requirements.txt
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.0
Werkzeug==1.0.1
apispec==4.0.0
marshmallow==3.8.0
apispec-webframeworks
flask-cors
Flask-Injector==0.12.3
googlemaps==2.5.1
rootpath==0.1.1
geopy==2.0.0
